+++ psycopg/psycopg/connection_async.py
@@ -288,7 +288,22 @@ class AsyncConnection(BaseConnection[Row]):
                 yield n
 
     async def wait(self, gen: PQGen[RV]) -> RV:
-        return await waiting.wait_async(gen, self.pgconn.socket)
+        try:
+            return await waiting.wait_async(gen, self.pgconn.socket)
+        except KeyboardInterrupt:
+            # TODO: this doesn't seem to work as it does for sync connections
+            # see tests/test_concurrency_async.py::test_ctrl_c
+            # In the test, the code doesn't reach this branch.
+
+            # On Ctrl-C, try to cancel the query in the server, otherwise
+            # otherwise the connection will be stuck in ACTIVE state
+            c = self.pgconn.get_cancel()
+            c.cancel()
+            try:
+                await waiting.wait_async(gen, self.pgconn.socket)
+            except e.QueryCanceled:
+                pass  # as expected
+            raise
 
     @classmethod
     async def _wait_conn(cls, gen: PQGenConn[RV], timeout: Optional[int]) -> RV: