+++ tests/test_concurrency.py
@@ -6,10 +6,12 @@ import os
 import sys
 import time
 import queue
-import pytest
+import signal
 import threading
 import subprocess as sp
 from typing import List
+
+import pytest
 
 import psycopg
 
@@ -201,3 +203,56 @@ def test_identify_closure(dsn):
     finally:
         conn.close()
         conn2.close()
+
+
+@pytest.mark.slow
+@pytest.mark.subprocess
+def test_ctrl_c(dsn):
+    if sys.platform == "win32":
+        sig = int(signal.CTRL_C_EVENT)
+        # Or pytest will recevie the Ctrl-C too
+        creationflags = sp.CREATE_NEW_PROCESS_GROUP
+    else:
+        sig = int(signal.SIGINT)
+        creationflags = 0
+
+    script = f"""\
+import os
+import time
+import psycopg
+from threading import Thread
+
+def tired_of_life():
+    time.sleep(1)
+    os.kill(os.getpid(), {sig!r})
+
+t = Thread(target=tired_of_life, daemon=True)
+t.start()
+
+with psycopg.connect({dsn!r}) as conn:
+    cur = conn.cursor()
+    ctrl_c = False
+    try:
+        cur.execute("select pg_sleep(2)")
+    except KeyboardInterrupt:
+        ctrl_c = True
+
+    assert ctrl_c, "ctrl-c not received"
+    assert (
+        conn.info.transaction_status == psycopg.pq.TransactionStatus.INERROR
+    ), f"transaction status: {{conn.info.transaction_status!r}}"
+
+    conn.rollback()
+    assert (
+        conn.info.transaction_status == psycopg.pq.TransactionStatus.IDLE
+    ), f"transaction status: {{conn.info.transaction_status!r}}"
+
+    cur.execute("select 1")
+    assert cur.fetchone() == (1,)
+"""
+    t0 = time.time()
+    proc = sp.Popen([sys.executable, "-s", "-c", script], creationflags=creationflags)
+    proc.communicate()
+    t = time.time() - t0
+    assert proc.returncode == 0
+    assert 1 < t < 2